Reorg RangeSelector code into file based on type of input
